Revert "Document that arguments after --help and --version are ignored"
This reverts commit bd431767a6d4bd5ef94f28f81f9bbff7eb4094d6.

I can't remember my full reasoning but I have
gradually decided against having this in the
help text. I think a big part was thinking
that even if ignoring further arguments is part
of the interface, it would be more confusing or
at least noisy to have it in the help text
(especially if tools parse docopt-style usage
for autocomplete/etc).